Move experiment functions to lib, keep ui elements in components
