+++ apps/class-solid/src/routes/index.tsx
@@ -1,11 +1,11 @@
 import { For, Show } from "solid-js";
 
 import { AnalysisCard, addAnalysis } from "~/components/Analysis";
-import { ExperimentCard, addExperiment } from "~/components/Experiment";
+import { ExperimentCard } from "~/components/Experiment";
 import { Button } from "~/components/ui/button";
 import { Flex } from "~/components/ui/flex";
 
-import { experiments } from "~/lib/store";
+import { addExperiment, experiments } from "~/lib/store";
 import { analyses } from "~/lib/store";
 
 export default function Home() {
